requirements.txt
# Google Agent Development Kit
google-adk

# Environment variable management
python-dotenv

# Additional dependencies (will be added as needed in later sprints)
